Don't babysit resources and methods
Checking for valid resource and a valid resource / method is silly,
checking every possibilities is a waste of time and adds unneeded
complexity to the code.
The only guarantee is that the provided wrappers work, if you
are doing requests on your own you need to know what you are doing.

Plus it removes so many lines it can't be wrong.