Add django test project
The project will use uwsgi.it uwsgi_it_api django application to
provide a local instance to make it easier to execute the tests.
Update tests to use as default a local instance of the test
project running by django runserver.

Since uwsgi_it_api is not published on pypi you need a local
checkout of uwsgi.it git repository for being able to install it.

Fix #3